requirements.txt
aiogram==3.4.1
python-dotenv==1.0.1
supabase==2.6.0
